Create improved chunking function
Instead of simply splitting the program string, which can only work for the
prelude and/or other programs whose "separation points" can be clearly
defined and known beforehand, we will add an additional instruction to
the CSE machine to pause and await further input. This way, we are able
to instead split the program into chunks from the JSON syntax tree
itself.

This will allow for arbitrary-length programs to be run. For now, we
will add the necessary chunking functions to the client app only.